Set version label based on tag if set
The `app.kubernetes.io/version` label is supposed to show the current
version of the software installed, but when the user overrides the
Jellyfin version using `image.tag` then the label becomes wrong since it
still uses `AppVersion`. By prioritising the `tag` over `AppVersion` for
this label, this is fixed.

Also update version of Helm chart. This is a bugfix since it makes sure
to show the correct version that should've been displayed anyway.

Signed-off-by: Florian Ruechel <code@inexplicity.de>